// public/model/enumeration/dnd5/DND5CreatureTypeEnum.ts
export enum DND5CreatureTypeEnum {
    ABBERATION = 'Abberation',
    BEAST = 'Beast',
    CELESTIAL = 'Celestial',
    CONSTRUCT = 'Construct',
    DRAGON = 'Dragon',
    ELEMENTAL = 'Elemental',
    FEY = 'Fey',
    FIEND = 'Fiend',
    GIANT = 'Giant',
    HUMANOID = 'Humanoid',
    MONSTROSITY = 'Monstrosity',
    OOZE = 'Ooze',
    PLANT = 'Plant',
    UNDEAD = 'Undead',
    NONE = ''
}